shim useWeb3React
we'll probably need to swap it out for web3auth -- this might make it a bit easier to not change lots of different places
