+++ src/App.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 
 import LuxonUtils from '@date-io/luxon';
-import { Web3Provider } from '@ethersproject/providers';
-import { Web3ReactProvider } from '@web3-react/core';
 import { SnackbarProvider } from 'notistack';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { BrowserRouter } from 'react-router-dom';
@@ -13,27 +11,12 @@ import { MuiPickersUtilsProvider } from '@material-ui/pickers';
 import { ThemeProvider } from '@material-ui/styles';
 
 import { ErrorBoundary, MainLayout } from 'components';
+import { Web3ReactProvider } from 'hooks/useWeb3React';
 import { createTheme } from 'theme';
 
 import { snackbarStyles, globalStyles } from './stitches.config';
 
 import './App.css';
-
-function getLibrary(provider: any): Web3Provider {
-  // This checks specifically whether the provider is
-  // an instance of a Web3Provider by checking the existence of this
-  // uniquely named method. Normally, we would want to use `instanceof`
-  // to check if a provider conforms to a specific class, but because
-  // some providers are injected into the window from other contexts,
-  // this check will fail, since those providers aren't derived from
-  // the same prototype tree.
-  const library =
-    typeof provider.jsonRpcFetchFunc !== 'undefined'
-      ? provider
-      : new Web3Provider(provider);
-  library.pollingInterval = 12000;
-  return library;
-}
 
 const theme = createTheme();
 
@@ -51,7 +34,7 @@ function App() {
           <QueryClientProvider client={queryClient}>
             <ThemeProvider theme={theme}>
               <MuiPickersUtilsProvider utils={LuxonUtils}>
-                <Web3ReactProvider getLibrary={getLibrary}>
+                <Web3ReactProvider>
                   <BrowserRouter>
                     <MainLayout />
                   </BrowserRouter>
