+++ src/components/MainLayout/getMainHeaderData.ts
@@ -1,9 +1,9 @@
-import { useWeb3React } from '@web3-react/core';
 import { client } from 'lib/gql/client';
 import { useQuery } from 'react-query';
 
 import { useAuthStep } from 'hooks/login';
 import useConnectedAddress from 'hooks/useConnectedAddress';
+import { useWeb3React } from 'hooks/useWeb3React';
 
 export const getMainHeaderData = (address: string, chainId: number) =>
   client.query(
