+++ src/utils/testing/index.tsx
@@ -1,7 +1,5 @@
 import { ReactElement, Suspense, useEffect } from 'react';
 
-import { Web3Provider } from '@ethersproject/providers';
-import { Web3ReactProvider, useWeb3React } from '@web3-react/core';
 import { NetworkConnector } from '@web3-react/network-connector';
 import { SnackbarProvider } from 'notistack';
 import { QueryClient, QueryClientProvider } from 'react-query';
@@ -10,6 +8,7 @@ import { RecoilRoot } from 'recoil';
 
 import { ThemeProvider } from '@material-ui/styles';
 
+import { Web3ReactProvider, useWeb3React } from 'hooks/useWeb3React';
 import { createTheme } from 'theme';
 
 import {
@@ -30,8 +29,6 @@ type TestWrapperProps = {
   queryClientCallback?: (queryClient: QueryClient) => void;
 };
 
-const defaultGetLibrary = (provider: any) => new Web3Provider(provider);
-
 const connector = new NetworkConnector({
   urls: { [chainId]: rpcUrl },
 });
@@ -47,7 +44,6 @@ const Web3Activator = ({ children, enabled }: Web3ActivatorProps) => {
 
 export const TestWrapper = ({
   children,
-  getLibrary = defaultGetLibrary,
   withWeb3 = false,
   queryClientCallback,
 }: TestWrapperProps) => {
@@ -65,7 +61,7 @@ export const TestWrapper = ({
     <RecoilRoot>
       <QueryClientProvider client={queryClient}>
         <SnackbarProvider>
-          <Web3ReactProvider getLibrary={getLibrary}>
+          <Web3ReactProvider>
             <Web3Activator enabled={withWeb3}>
               <MemoryRouter>
                 <ThemeProvider theme={theme}>
